Improve thread safety
Win32 C++ later::later was not working if the first
call to later::later (in either R or C++) was performed
on a different thread. This is because some of the
machinery we use (message-only window) is lazily
instantiated, and if it gets created on a background
thread then things won't go well. This was probably
a problem on posix as well (addInputHandler is probably
not threadsafe) but it was much more obvious on win32
as the repro case happens 100%.

This fix causes those lazily loaded structures to be
eagerly loaded, in package .onLoad(). It also changes
the win32 C-flavored doExecLater to detect if it's
being called from a different thread, and doesn't
directly call (the non-threadsafe) SetTimer() in that
case, but rather sends a user message to tell the main
thread to call SetTimer().