Fix #4: later breaks socket wait (and `future` package)
Previously, our POSIX input handler would become active (the
associated file descriptor becomes "hot" i.e. ready to read)
the moment a callback was scheduled for the future. This means
our input handler would be called constantly by R. We lived
with this to keep the implementation simple, thinking the
input handler executes so quickly that the time spent didn't
matter.

That may still be true, but a bigger problem arose when I
tried to use the future package with promises. When using
future's multisession mode, futures use sockets for IPC with
worker processes. These sockets would hang if our input
handler was active, due to code in R_SocketWait that insists
that no other input handlers be active in order for socket
I/O to proceed. I'm not sure if that's reasonable behavior
on R_SocketWait's part, but that code is 17 years old so I'm
not about to suggest changes now.

I fixed the issue by taking the complexity hit and writing a
pthread-based Timer class (for POSIX only; Window's WM_TIMER
serves a similar purpose but far more elegantly). With the
Timer class, we're able to schedule the file descriptor to
go hot at arbitrary times in the future, letting us keep it
cold for the vast majority of times.